BookStoreBackend/requirements.txt
Django==5.1.2
djangorestframework==3.15.1
pandas==2.2.3
PyJWT==2.8.0
PyJWT==2.10.1
PyMySQL==1.1.1
